rename file for Side
